Simplify Twitter operator spec (#21)
The previous version of this document was overprescribed. The removed
entries were either unnecessary, or have been in flux. In any case, a
lesson is learned here: specify only those things that we know for sure
are the way something should be done. When in doubt, leave it to the
discretion of the role owner.